# Optimisations for get and memoize