Optimization of `get`, `caching` and `memoize`

* Hand-optimise `CacheKeyBuilder` (used by both `get` and `memoize`) to use a loop and a StringBuilder
* Hand-optimise `MethodCallToStringConverter` (used by `memoize`) similarly
* Special treatment for `memoize` because we know in this case that the key is a single string, not an `Any*`
* Special handling in `_caching` for already-completed futures (such as those returned by the in-memory cache implementations). If the future is already completed, we can avoid calling `recover` and `flatMap` on it, which means we avoid spawning more futures. This makes a HUGE difference. This one optimisation accounts for the vast majority of the speedup.

## Before

```
[info] Benchmark                                      Mode  Cnt      Score      Error  Units
[info] CaffeineBenchmark.caffeineGet                  avgt   30     20.547 ±    0.596  ns/op
[info] CaffeineBenchmark.scalacacheGetNoMemoize       avgt   30    334.866 ±   32.630  ns/op
[info] CaffeineBenchmark.scalacacheGetTypedNoMemoize  avgt   30    340.825 ±   36.284  ns/op
[info] CaffeineBenchmark.scalacacheGetWithMemoize     avgt   30  39801.345 ± 2974.877  ns/op
```

## After

```
[info] Benchmark                                      Mode  Cnt    Score     Error  Units
[info] CaffeineBenchmark.caffeineGet                  avgt   30   20.110 ±   0.270  ns/op
[info] CaffeineBenchmark.scalacacheGetNoMemoize       avgt   30  198.882 ±  10.764  ns/op
[info] CaffeineBenchmark.scalacacheGetTypedNoMemoize  avgt   30  206.160 ±  12.344  ns/op
[info] CaffeineBenchmark.scalacacheGetWithMemoize     avgt   30  580.810 ± 129.729  ns/op
```

Not bad for an afternoon's work 😄  

/cc @wuservices @ben-manes in case you're interested